+++ utils/btcprice.go
@@ -9,12 +9,12 @@ import (
 const blockchainTickerURL = "https://blockchain.info/ticker"
 
 // BitcoinPrice returns the lastest bitcoin price in Penny (1USD = 100Penny)
-func BitcoinPrice() (_ int, err error) {
+func BitcoinPrice() (_ int64, err error) {
 	_, body, _ := gorequest.New().Get(blockchainTickerURL).EndBytes()
 	return bitcoinPriceWithByteFromBlockchain(body)
 }
 
-func bitcoinPriceWithByteFromBlockchain(data []byte) (int, error) {
+func bitcoinPriceWithByteFromBlockchain(data []byte) (int64, error) {
 	m := map[string]struct { // use pointer so it fails fast
 		Last float64 `json:"last"`
 	}{}
@@ -24,5 +24,5 @@ func bitcoinPriceWithByteFromBlockchain(data []byte) (int, error) {
 		return 0, err
 	}
 
-	return int(100 * m["USD"].Last), nil
+	return int64(100 * m["USD"].Last), nil
 }
